services/properties/requirements.txt
Flask==1.0.3
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.3.2
psycopg2-binary==2.7.7